Lifted ETag.EntityTag to org.http4s package
