+++ core/src/main/scala/org/http4s/headers/ETag.scala
@@ -17,37 +17,24 @@
 package org.http4s
 package headers
 
+import org.http4s
+import org.http4s.EntityTag.{Strong, Weakness}
 import org.http4s.parser.HttpHeaderParser
 import org.http4s.util.Writer
-import cats.Show
-import org.http4s.headers.ETag.EntityTag.{Strong, Weakness}
 
 object ETag extends HeaderKey.Internal[ETag] with HeaderKey.Singleton {
-  final case class EntityTag(tag: String, weakness: Weakness = Strong) {
-    override def toString(): String =
-      weakness match {
-        case EntityTag.Weak => "W/\"" + tag + '"'
-        case EntityTag.Strong => "\"" + tag + '"'
-      }
-  }
 
-  object EntityTag {
-    implicit val http4sShowForEntityTag: Show[EntityTag] =
-      Show.fromToString
-
-    sealed trait Weakness extends Product with Serializable
-    case object Weak extends Weakness
-    case object Strong extends Weakness
-  }
+  type EntityTag = http4s.EntityTag
+  val EntityTag: http4s.EntityTag.type = http4s.EntityTag
 
   def apply(tag: String, weakness: Weakness = Strong): ETag =
-    ETag(EntityTag(tag, weakness))
+    ETag(http4s.EntityTag(tag, weakness))
 
   override def parse(s: String): ParseResult[ETag] =
     HttpHeaderParser.ETAG(s)
 }
 
-final case class ETag(tag: ETag.EntityTag) extends Header.Parsed {
+final case class ETag(tag: EntityTag) extends Header.Parsed {
   def key: ETag.type = ETag
   override def value: String = tag.toString()
   override def renderValue(writer: Writer): writer.type = writer.append(value)
