+++ core/src/main/scala/org/http4s/headers/If-Match.scala
@@ -27,7 +27,7 @@ object `If-Match` extends HeaderKey.Internal[`If-Match`] with HeaderKey.Singleto
   /** Match any existing entity */
   val `*` = `If-Match`(None)
 
-  def apply(first: ETag.EntityTag, rest: ETag.EntityTag*): `If-Match` =
+  def apply(first: EntityTag, rest: EntityTag*): `If-Match` =
     `If-Match`(Some(NonEmptyList.of(first, rest: _*)))
 
   override def parse(s: String): ParseResult[`If-Match`] =
@@ -39,7 +39,7 @@ object `If-Match` extends HeaderKey.Internal[`If-Match`] with HeaderKey.Singleto
   *
   * [[https://tools.ietf.org/html/rfc7232#section-3.1 RFC-7232 Section 3.1]]
   */
-final case class `If-Match`(tags: Option[NonEmptyList[ETag.EntityTag]]) extends Header.Parsed {
+final case class `If-Match`(tags: Option[NonEmptyList[EntityTag]]) extends Header.Parsed {
   override def key: `If-Match`.type = `If-Match`
   override def value: String =
     tags match {
