+++ core/src/main/scala/org/http4s/headers/If-None-Match.scala
@@ -37,14 +37,14 @@ object `If-None-Match` extends HeaderKey.Internal[`If-None-Match`] with HeaderKe
   /** Match any existing entity */
   val `*` = `If-None-Match`(None)
 
-  def apply(first: ETag.EntityTag, rest: ETag.EntityTag*): `If-None-Match` =
+  def apply(first: EntityTag, rest: EntityTag*): `If-None-Match` =
     `If-None-Match`(Some(NonEmptyList.of(first, rest: _*)))
 
   override def parse(s: String): ParseResult[`If-None-Match`] =
     HttpHeaderParser.IF_NONE_MATCH(s)
 }
 
-final case class `If-None-Match`(tags: Option[NonEmptyList[ETag.EntityTag]]) extends Header.Parsed {
+final case class `If-None-Match`(tags: Option[NonEmptyList[EntityTag]]) extends Header.Parsed {
   override def key: `If-None-Match`.type = `If-None-Match`
   override def value: String =
     tags match {
